// src/app/model/agregados.ts
export class Agregados {
  id: any;
  logado: boolean;
  deslogado: boolean;
  listaGeral: any;
  listaUpgrade: any;
  listaDiaria: any;
  descricao: any;
  total: any;
  promocao: any;
  categoria; any;
  calcularSeguro = 0;


  gps = 0;
  cadeiraBebe = 0;
  assentoEleva = 0;
  cadeiraBebeConforto = 0;
  percentSeguroBasico: any;
  calcularSeguroBasico: any;
  seguroNormalBasico: any;
  percentSeguroPlus: any;
  calcularSeguroPlus: any;
  seguroplanoCompleto: any;
  seguroNormalCompleto: any;
  seguroplanoBasico: any;
  seguroAtivoPlano = false;
  seguroAtivoDiaria = false;
  calcularDiariaPromocao: any;
  seguroPercentual: any;

  valorGrupo: any;
  nomeGrupo: any;
  diariaContrata: any;

}